refactor(git): extract ref helpers
